test(lifecycle): split the query commands into their own file
The added assertions took lifecycle.rs to 602 code lines, past the 500 hard
limit, and line-limit went red on the pull request. I should have measured this
before pushing — I checked it on the watch file earlier in the day and did not
repeat it here.

The cut is the file's own "Query" section marker: up/down/restart stay in
lifecycle.rs, and the commands that read back what those did move to
lifecycle_query.rs. 233 and 370 code lines.

Refs #1180
